requirements/local.txt
-r base.txt

psycopg2-binary==2.9.10  # https://github.com/psycopg/psycopg2
watchfiles==1.1.1  # https://github.com/samuelcolvin/watchfiles
Werkzeug[watchdog]==3.1.3 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb

# Testing
# ------------------------------------------------------------------------------

mypy==1.18.2  # https://github.com/python/mypy
django-stubs==4.2.7  # https://github.com/typeddjango/django-stubs
pytest==8.4.2  # https://github.com/pytest-dev/pytest
pytest-cov==7.0.0  # https://github.com/pytest-dev/pytest-cov
pytest-sugar==1.1.1  # https://github.com/Frozenball/pytest-sugar
pytest-xdist==3.8.0  # https://github.com/pytest-dev/pytest-xdist (parallel test execution)
djangorestframework-stubs==1.8.0  # https://github.com/typeddjango/djangorestframework-stubs
responses==0.25.8  # https://github.com/getsentry/responses
vcrpy==7.0.0

# Profiling
# ------------------------------------------------------------------------------
django-cprofile-middleware
django-silk==5.4.0

# Code quality
# ------------------------------------------------------------------------------
flake8==7.3.0 # https://github.com/PyCQA/flake8
flake8-isort==7.0.0  # https://github.com/gforcada/flake8-isort
coverage==7.11.0  # https://github.com/nedbat/coveragepy
black==25.9.0  # https://github.com/psf/black
pylint-django==2.6.1  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==4.5.1  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.3  # https://github.com/FactoryBoy/factory_boy
django-debug-toolbar==4.4.6  # https://github.com/jazzband/django-debug-toolbar
django-extensions==4.1  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.2.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.11.1  # https://github.com/pytest-dev/pytest-django

# PDFs
# ------------------------------------------------------------------------------
pypdf

# Office files
# ------------------------------------------------------------------------------
python-docx
openpyxl
python-pptx

# Not directly required, pinned by Snyk to avoid a vulnerability
# ------------------------------------------------------------------------------
django==4.2.24  # pyup: < 4.0  # https://www.djangoproject.com/
twisted>=24.7.0rc1 # not directly required, pinned by Snyk to avoid a vulnerability
ipython>=8.10.0 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=70.0.0 # not directly required, pinned by Snyk to avoid a vulnerability
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
tornado>=6.4.2 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
pillow>=10.3.0 # not directly required, pinned by Snyk to avoid a vulnerability
